refactor: replace schema-diff + upgrade-api-version with connect:sf-api-upgrade
Consolidates the two commands into a single connect:sf-api-upgrade command.
Running without --confirm shows the schema diff; passing --confirm <connection-name>
shows the diff and then runs the upgrade in one step. Removes --force flag —
the backend owns the decision on whether an upgrade can proceed.
